git-svn-id: http://svn.osgeo.org/fdo/trunk@2510 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Utilities/SchemaMgr/Src/Sm/Ph/DependencyWriter.cpp
@@ -99,7 +99,7 @@ void FdoSmPhDependencyWriter::SetOrderType(FdoStringP sValue)
 
 void FdoSmPhDependencyWriter::SetCardinality(long lValue)
 {
-	SetDouble(L"",L"fkcardinality", lValue);
+	SetLong(L"",L"fkcardinality", lValue);
 }
 
 void FdoSmPhDependencyWriter::Modify( FdoStringP pkTableName, FdoStringP fkTableName )
